Moved to more directory based build system - heading towards CMAKE
