+++ .gitignore
@@ -4,3 +4,6 @@
 CUDAProb3/
 Prob3plusplus/
 ProbGPU/
+
+Bin/
+Lib/
