--- OscillatorConstants.h
@@ -1,13 +1,0 @@
-#ifndef __OSCILLATOR_CONSTANTS__
-#define __OSCILLATOR_CONSTANTS__
-
-/**
- * @file OscillatorConstants.h
- */
-
-/**
- * @brief Different verbosity levels for console output
- */
-enum Verbosity{NONE,INFO};
-
-#endif
